Final version of crazy dotfiles2appdata.sh script
This is what has been used for the last few months, but not committed until now.
Will immediately delete and replace with new more sensible script which only
does relevant things on the WSL side, as now this repository is to be cloned on
Windows and symlinked using Windows scripts.